[refactor] fix splint bugs

+++ config.def.h
@@ -7,26 +7,26 @@
 
 /* colors */
 static const uint16_t dir_b     = TB_DEFAULT;
-static const uint16_t dir_f     = 33;
+static const uint16_t dir_f     = (uint16_t)33;
 static const uint16_t file_b    = TB_DEFAULT;
 static const uint16_t file_f    = TB_DEFAULT;
 static const uint16_t frame_b   = TB_DEFAULT;
 static const uint16_t frame_f   = TB_DEFAULT;
 static const uint16_t other_b   = TB_DEFAULT;
-static const uint16_t other_f   = 3;
+static const uint16_t other_f   = (uint16_t)3;
 static const uint16_t pane_l_b  = TB_DEFAULT;
-static const uint16_t pane_l_f  = 166;
+static const uint16_t pane_l_f  = (uint16_t)166;
 static const uint16_t pane_r_b  = TB_DEFAULT;
-static const uint16_t pane_r_f  = 5;
-static const uint16_t search_b  = 166;
-static const uint16_t search_f  = 255;
-static const uint16_t status_b  = 234;
+static const uint16_t pane_r_f  = (uint16_t)5;
+static const uint16_t search_b  = (uint16_t)166;
+static const uint16_t search_f  = (uint16_t)255;
+static const uint16_t status_b  = (uint16_t)234;
 static const uint16_t status_f  = TB_DEFAULT;
-static const uint16_t sprompt_f = 33;
-static const uint16_t sprompt_b = 234;
-static const uint16_t serr_f    = 124;
-static const uint16_t serr_b    = 234;
-static const uint16_t exec_f    = 2;
+static const uint16_t sprompt_f = (uint16_t)33;
+static const uint16_t sprompt_b = (uint16_t)234;
+static const uint16_t serr_f    = (uint16_t)124;
+static const uint16_t serr_b    = (uint16_t)234;
+static const uint16_t exec_f    = (uint16_t)2;
 
 /* bookmarks */
 static Bookmark bmarks[] = {
